13/01/2016 Se termina el borrar (de momento)

+++ application/controllers/Articulos.php
@@ -1,24 +1,42 @@
-<?php defined('BASEPATH') OR exit('No direct script access allowed');
+<?php
 
-class Articulos extends CI_Controller
-{
+defined('BASEPATH') OR exit('No direct script access allowed');
 
-    public function index()
-    {
+class Articulos extends CI_Controller {
+
+    public function index() {
         $data['filas'] = $this->db->query('select * from articulos ')->result_array();
         $this->load->view('articulos/index', $data);
     }
 
-    public function borrar($id)
-    {
-        $data['id'] = $id;
-        $this->load->view('articulos/borrar',$data);
+    public function borrar($id = NULL) {
+        if ($this->input->post('borrar') !== NULL) {
+            $id = $this->input->post('id');
+            if ($id !== NULL) {
+                $res = $this->db->query('delete from articulos where id = ?', array($id));
+            }
+            redirect('articulos/index');
+        }
+        else 
+        {
+            if ($id === NULL)
+            {
+                redirect('articulos/index');
+            } 
+            else 
+            {
+                $res = $this->db->query('select * from articulos where id =?', array($id));
+                if ($res->num_rows() === 0){
+                    redirect('articulos/index');
+                }
+                else 
+                {
+                    $data= $res->row_array();
+                    $this->load->view('articulos/borrar', $data);
+                }
+                
+            }
+        }
     }
-    
-    public function hacer_borrado()
-    {
-        $id = $this->input->post('id');
-        $res = $this->db->query('delete from articulos where id = ?', array($id));
-        redirect('articulos/index');
-    }
+
 }
